+++ src/components/UserPage/VestingContractsSection/index.ts
@@ -1,0 +1,1 @@
+export { default as VestingContractsSection } from './VestingContractsSection';
