+++ src/components/UserPage/index.ts
@@ -1,0 +1,2 @@
+export * from './styled';
+export * from './VestingContractsSection';
