+++ src/pages/User.tsx
@@ -1,9 +1,7 @@
 import { observer } from 'mobx-react';
 import { useHistory } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
-import styled from 'styled-components';
 import moment from 'moment';
-import { useState, useEffect } from 'react';
 import {
   BlockchainLink,
   Row,
@@ -12,28 +10,14 @@ import {
   Subtitle,
   Button,
 } from '../components/common';
-import UserVestingInfoModal from '../components/UserVestingInfoModal';
-import useContract from '../hooks/useContract';
-import { formatBalance, bnum } from '../utils';
+import { formatBalance } from '../utils';
 import { useContext } from '../contexts';
 import useExporters from '../hooks/useExporters';
-import ERC20Json from '../contracts/ERC20.json';
-
-const TitleRow = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const ListRow = styled.div`
-  display: flex;
-  alignitems: center;
-  padding: 6px 0px;
-  borderbottom: ${({ borderBottom }) => (borderBottom ? '1px solid' : '')};
-  cursor: ${({ clickable }) => (clickable ? 'pointer' : 'default')};
-  &:hover {
-    opacity: ${({ clickable }) => (clickable ? '0.5' : '1')};
-  }
-`;
+import {
+  TitleRow,
+  ListRow,
+  VestingContractsSection,
+} from 'components/UserPage';
 
 const UserPage = observer(() => {
   let history = useHistory();
@@ -47,11 +31,6 @@ const UserPage = observer(() => {
   const userInfo = daoStore.getUser(userAddress);
   const networkName = configStore.getActiveChainName();
   const redeemsLeft = daoStore.getUserRedeemsLeft(userAddress);
-  const userVestingContracts = daoStore.getUserVestingContracts(userAddress);
-  const [isVestingInfoModalOpen, setIsVestingInfoModalOpen] = useState(false);
-  const [selectedModalVestingContract, setSelectedModalVestingContract] =
-    useState(null);
-  const [tokenVestingContracts, setTokenVestingContracts] = useState([]);
 
   const getExportFileName = () => {
     return `history_${userAddress}-${moment().format('YYYY-MM-DD')}`;
@@ -68,33 +47,6 @@ const UserPage = observer(() => {
     const csvString = await exportToCSV(historyItems);
     triggerDownload(csvString, `${getExportFileName()}.csv`, 'text/csv');
   };
-
-  const DXD = useContract(
-    configStore.getNetworkContracts().votingMachines.dxd.token,
-    ERC20Json.abi
-  );
-
-  const updateUserVestingContracts = () => {
-    const contracts = Promise.all(
-      userVestingContracts.map(async contract => {
-        try {
-          const balance = await DXD?.balanceOf(contract.address);
-          return {
-            ...contract,
-            value: bnum(balance),
-          };
-        } catch (e) {
-          return {
-            ...contract,
-            value: bnum('0'),
-          };
-        }
-      })
-    );
-    contracts.then(setTokenVestingContracts);
-  };
-
-  useEffect(updateUserVestingContracts, []); // eslint-disable-line
 
   return (
     <Box>
@@ -188,44 +140,7 @@ const UserPage = observer(() => {
           </ListRow>
         );
       })}
-
-      <TitleRow>
-        <h2>Vesting Contracts</h2>
-      </TitleRow>
-
-      {tokenVestingContracts.length ? (
-        tokenVestingContracts.map((contract, i, arr) => {
-          return (
-            <ListRow
-              clickable
-              key={contract.address}
-              borderBottom={i < arr.length - 1}
-              onClick={() => {
-                if (!contract.address) return;
-                setSelectedModalVestingContract(contract);
-                setIsVestingInfoModalOpen(true);
-              }}
-            >
-              <span>
-                {contract.address} / Cliff:{' '}
-                {moment.unix(Number(contract.cliff)).format('LL')}/ Value:{' '}
-                {formatBalance(contract.value)} DXD
-              </span>
-            </ListRow>
-          );
-        })
-      ) : (
-        <ListRow> - No Vesting Contracts found for {userAddress} - </ListRow>
-      )}
-      <UserVestingInfoModal
-        contract={selectedModalVestingContract}
-        isOpen={isVestingInfoModalOpen}
-        onDismiss={() => {
-          setSelectedModalVestingContract(null);
-          setIsVestingInfoModalOpen(false);
-        }}
-        updateContracts={updateUserVestingContracts}
-      />
+      <VestingContractsSection userAddress={userAddress} />
     </Box>
   );
 });
